Module giga::editor

Where the magic happens.

The editor module contains the main structure of the program: the Editor. The program is architected around an MVC-like pattern. The Editor is the controller, it contains the model (View + Some metadatas) and the view is handled by an implementation of the TerminalDrawer trait.

Asynchronous architecture

There are three threads in the program:

Main thread, responsible for handling user input and modifying the view

Git thread, responsible for computing the diff between the current file and the current commit and updating the diff field of the editor

The tui thread, responsible for drawing the view to the terminal. This thread listens to both the Main thread and the Git thread (using message passing) and redraws the view when needed.

The View

The view (View module) represents the portion of the file that is currently displayed on and it owns the representation of the file (File module) in memory. Most of the tasks performed by the editor act on the view (moving the cursor, scrolling, inserting/deleting characters, etc…). The view produces a syntax highlighted representation of the file. The logic for coloring the file is handled in the Colorizer struct (in the color module) and is performed synchronously by the main thread in the file module.

Command system

To act on the editor, the user inputs a char to stdin. Which is then parsed into a Command enum variant depending on the current mode of the editor. This logic is handled by the Command module. The Command is then executed by the Editor (Editor::execute method) which produces a RefreshOrder enum variant that represents which portion of the screen needs to be redrawn. The tui thread (Editor::init_tui_thread method) then uses this information to redraw the screen.

Drawing system

The drawing system is handled by the TerminalDrawer trait. Which exists to abstract the drawing system from the rest of the program (for modularity and mocking purposes). The TerminalDrawer trait is implemented in the terminal module. The current implementation uses the termion crate to draw to the terminal.

Git integration

If the current file is in a git repository, the editor will display the current branch/commit/tag in the status bar and will display the diff between the current file and the current commit in the left margin. This is done by the git module.

This task is performed asynchronously by the git thread (Editor::init_git_thread method). The git thread computes the diff between the current file and the current commit and stores it in the diff field. It then sends a unit signal to the tui thread to notify it that the diff has changed.

Resizing the terminal

To allow the user to resize the terminal, the tui thread listens to the SIGWINCH signal and sends a unit signal to the tui thread to notify it that the terminal has been resized. This logic is handled by the signal module, which is called in the (Editor::init_resize_listener method).

Modules

command 🔒

Parse user input into editor commands

signal 🔒

Intercepting terminal resize events

terminal 🔒

Provides the terminal drawing logic

view 🔒

Portion of the file being displayed

Macros

arc_mutex 🔒

Macro to create arc mutexes quickly

Structs

Editor

Editor structure represents the state of the program

Enums

Mode

Mode of the editor

RefreshOrder