Enum giga::editor::command::Command

pub enum Command {
    Quit,
    Move(isize, isize),
    Save,
    Rename(Option<char>),
    ToggleMode,
    ToggleRename,
    Insert(char),
    Delete,
    InsertNewLine,
    CommandBlock(Vec<Command>),
}

Commands that can be executed by the editor

Variants

Quit

Quit the editor

Move(isize, isize)

Move the cursor by dx, dy

Save

Save the file

Rename(Option<char>)

Rename the file

ToggleMode

Toggle mode

ToggleRename

Toggle rename

Insert(char)

Insert a character

Delete

Delete a character

InsertNewLine

Insert a new line

CommandBlock(Vec<Command>)

CommandBlock

Implementations

impl Command

pub fn parse(key: Key, mode: &Mode) -> Result<Self, &'static str>

Parse a command from a termion::event::Key object

fn parse_normal_mode(key: Key) -> Result<Self, &'static str>

Parse a command in normal mode from a termion::event::Key object

fn parse_insert_mode(key: Key) -> Result<Self, &'static str>

Parse a command in insert mode from a termion::event::Key object

fn parse_rename_mode(key: Key) -> Result<Self, &'static str>

fn parse_insert_mode_char(c: char) -> Result<Self, &'static str>

Parse a character in insert mode