Struct giga::editor::Editor

pub struct Editor {
    file_path: String,
    file_name: Arc<Mutex<String>>,
    view: Arc<Mutex<View>>,
    mode: Arc<Mutex<Mode>>,
    git_ref: Arc<Mutex<Option<String>>>,
    pub diff: Arc<Mutex<Option<Vec<Patch>>>>,
}

Editor structure represents the state of the program

Fields
file_path: String

The path of the file being edited

file_name: Arc<Mutex<String>>

The name of the file being edited

view: Arc<Mutex<View>>

The current view of the file

mode: Arc<Mutex<Mode>>

The mode of the editor

git_ref: Arc<Mutex<Option<String>>>

Git Branch/Commit/Tag if any

diff: Arc<Mutex<Option<Vec<Patch>>>>

Git diff since last commit if any

Implementations

impl Editor

pub fn new(file_path: &str) -> Self

Create a new editor

pub fn open(path: &str) -> Result<Self, Error>

Open a file in the editor

fn split_path_name(path: &str) -> (String, String, String)

Split a file path into a relative path and a name

fn save(&self)

Save the current file

fn rename(&mut self, c: Option<char>)

Rename the current file

fn execute(&mut self, cmd: Command) -> RefreshOrder

Execute an editor command

Quit: exit the program

Move: move the cursor

Save: save the file

Rename: rename the file

ToggleMode: toogle editor mode

Insert: insert a character

Delete: delete a character

fn toggle_mode(&mut self)

Toggle the mode of the editor between normal and insert

fn toggle_rename(&mut self)

fn init_git_thread(&mut self, sender: Sender<RefreshOrder>)

Initialize git operations

fn get_status_bar_infos( mode: &Arc<Mutex<Mode>>, file_name: &Arc<Mutex<String>>, git_ref: &Arc<Mutex<Option<String>>> ) -> StatusBarInfos

Get the status bar infos

fn refresh_tui( tui: &mut TermionTerminalDrawer, view: &mut View, diff: &Arc<Mutex<Option<Vec<Patch>>>>, status_bar_infos: &StatusBarInfos, refresh_order: RefreshOrder )

Refresh the TUI

fn init_tui_thread(&mut self, refresh_receiver: Receiver<RefreshOrder>)

Initialize the tui drawing thread

This thread listens from three channels:

cmd_rx: commands refresh orders from the main thread

diff_changed: a channel that is notified when the diff has changed

resize_rx: a channel that is notified when the terminal has been resized It then draws the TUI accordingly.

pub fn run(&mut self)

Run the editor loop