Struct giga::editor::view::View

pub struct View {
    file: File,
    pub start_line: usize,
    pub start_col: usize,
    pub height: usize,
    pub width: usize,
    pub cursor: (usize, usize),
}

The View struct represents the actual portion of the File being displayed.

Fields
file: File

The file being displayed

start_line: usize

The line number of the first line being displayed

start_col: usize

The column number of the first column being displayed

height: usize

The number of lines being displayed

width: usize

The number of columns being displayed

cursor: (usize, usize)

The position of the cursor in the view

Implementations

impl View

fn navigate_y(&mut self, dy: isize) -> bool

Navigate along the y axis and eventually scroll the view

fn navigate_x(&mut self, dx: isize) -> bool

Navigate along the x axis and eventually scroll the view