Struct giga::tui::Tui

pub struct Tui {
    pub stdout: RawTerminal<Stdout>,
}

Terminal User Interface Responsible for drawing the editor and handling user input using termion in raw mode

Fields
stdout: RawTerminal<Stdout>
Implementations

impl Tui

pub fn new() -> Self

pub fn get_term_size(&self) -> (u16, u16)

Get the terminal size Returns a tuple of (width, height)

pub fn clear(&mut self)

Clear the screen

pub fn cleanup(&mut self)

Cleanup the terminal before exiting the program This will :

Flush the stdout buffer

Clear the screen

Move the cursor to the top left

Reset the terminal colors

Reset the terminal cursor

Disable raw mode

Show the cursor

pub fn draw_status_bar(&mut self, status_bar: &StatusBar, height: u16, width: u16)

Draw the status bar The status bar is displayed at the bottom of the screen It contains the current mode and the file name

pub fn draw_line_numbers(&mut self, line: usize)

Draw the line numbers The line numbers are displayed at the left of the screen in blue

pub fn draw_view(&mut self, view: &View, sb: &StatusBar)

Draw the view on the screen The view is the portion of the file being displayed

pub fn refresh_lines(&mut self, view: &View, lines: HashSet<u16>)

pub fn move_cursor(&mut self, x: usize, y: usize)